backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
sqlalchemy
python-dotenv
requests
openai
pydantic
